fix(LINT_LICENSING): lint and licensing issue resolved
UIC-1001
